Fixed windows build
